Include RN SDK in Expo package
