+++ packages/expo/app.plugin.js
@@ -1,1 +1,1 @@
-module.exports = require('./build');
+module.exports = require('./build/plugin');
