+++ packages/expo/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@datadog/mobile-expo",
-  "version": "1.0.0-rc9",
+  "name": "expo-datadog",
+  "version": "45.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "build/index.js",
   "scripts": {
     "build": "expo-module build",
     "clean": "expo-module clean",
@@ -10,6 +10,9 @@
     "prepare": "expo-module prepare",
     "prepublishOnly": "expo-module prepublishOnly",
     "expo-module": "expo-module"
+  },
+  "dependencies": {
+    "@datadog/mobile-react-native": "1.0.0-rc9"
   },
   "author": "",
   "license": "ISC"
