+++ packages/expo/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "expo-module-scripts/tsconfig.base",
+  "extends": "expo-module-scripts/tsconfig.plugin",
   "compilerOptions": {
     "outDir": "./build",
     "rootDir": "./src"
